requirements.txt
suntime>=1.3.2
smartcoop-python-sdk>=1.1.7